requirements.txt
requests
pyyaml
fire
